Move some ops to array_ops.
